Added the BigDecimal to fruit Dir.

+++ shop/src/main/java/com/burmau/shop/model/fruit/Fruit.java
@@ -1,8 +1,10 @@
 package com.burmau.shop.model.fruit;
 
+import java.math.BigDecimal;
+
 public abstract class Fruit {
     abstract String getFruitType();
-    abstract double getFruitPrice();
+    abstract BigDecimal getFruitPrice();
     abstract String getFruitDescription();
 
 }
